+++ tapo/src/api/hub_handler.rs
@@ -43,6 +43,7 @@ impl HubHandler {
     /// Returns *child device list* as [`ChildDeviceHubResult`].
     /// It is not guaranteed to contain all the properties returned from the Tapo API
     /// or to support all the possible devices connected to the hub.
+    /// If the deserialization fails, or if a property that you care about it's not present, try [`HubHandler::get_child_device_list_json`].
     pub async fn get_child_device_list(&self) -> Result<Vec<ChildDeviceHubResult>, Error> {
         self.client
             .get_child_device_list::<ChildDeviceListHubResult>()