+++ tapo/src/responses/child_device_list_hub_result/t100_result.rs
@@ -5,36 +5,52 @@ use crate::responses::{decode_value, DecodableResultExt, Status, TapoResponseExt
 
 /// T100 motion sensor.
 ///
-/// Specific properties: `detected`.
+/// Specific properties: `detected`, `report_interval`,
+/// `last_onboarding_timestamp`, `status_follow_edge`.
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
 #[allow(missing_docs)]
 pub struct T100Result {
+    // Common properties to all Hub child devices.
     pub at_low_battery: bool,
     pub avatar: String,
     pub bind_count: u32,
     pub category: String,
-    pub detected: bool,
     pub device_id: String,
     pub fw_ver: String,
     pub hw_id: String,
     pub hw_ver: String,
     pub jamming_rssi: i16,
     pub jamming_signal_level: u8,
-    #[serde(rename = "lastOnboardingTimestamp")]
-    pub last_onboarding_timestamp: u64,
     pub mac: String,
     pub nickname: String,
     pub oem_id: String,
     pub parent_device_id: String,
     pub region: String,
-    /// The time in seconds between each report.
-    pub report_interval: u32,
     pub rssi: i16,
     pub signal_level: u8,
     pub specs: String,
-    pub status_follow_edge: bool,
     pub status: Status,
     pub r#type: String,
+    // Specific properties to this device.
+    pub detected: bool,
+    #[serde(rename = "lastOnboardingTimestamp")]
+    pub last_onboarding_timestamp: u64,
+    /// The time in seconds between each report.
+    pub report_interval: u32,
+    pub status_follow_edge: bool,
+}
+
+#[cfg(feature = "python")]
+#[pyo3::pymethods]
+impl T100Result {
+    /// Gets all the properties of this result as a dictionary.
+    pub fn to_dict(&self, py: pyo3::Python) -> pyo3::PyResult<pyo3::Py<pyo3::types::PyDict>> {
+        let value = serde_json::to_value(self)
+            .map_err(|e| pyo3::exceptions::PyException::new_err(e.to_string()))?;
+
+        crate::python::serde_object_to_py_dict(py, &value)
+    }
 }
 
 impl TapoResponseExt for T100Result {}